requirements.txt
python-dotenv
langchain
huggingface_hub


